changed type of getDistance function
changed type and return type of getDistance function from double to uint32_t

+++ main.c
@@ -81,7 +81,7 @@ void portB_segments_init(void){
 
 
 }
-double getDistance(double p1[],double p2[]){
+uint32_t getDistance(double p1[],double p2[]){
    double lat1=p1[0];
    double lon1=p1[1];
    double lat2=p2[0];
@@ -96,7 +96,7 @@ double getDistance(double p1[],double p2[]){
    double c=2*atan2(sqrt(a),sqrt((1-a)));
    double d=(R*c);
 
-     return d;
+     return (uint32_t)d;
 }
